Rename hanami-esbuild-plugin -> esbuild-plugin
Go with a shorter name for tidiness
